Add open shader menu
